chore(build): move worker configs to vite/worker-configs.{esm,umd}.js
Extract apidomWorkerConfig and editorWorkerConfig out of vite.config.esm.js
and vite.config.umd.js into dedicated files. Update build scripts to import
mainConfig and worker configs from their respective locations.

vite.config.{esm,umd}.js now contain only the main bundle config.
