+++ vite/worker-configs.umd.js
@@ -1,0 +1,54 @@
+import { defineConfig } from 'vite';
+import { resolve, dirname } from 'path';
+import { fileURLToPath } from 'url';
+
+import { logger, sharedOnwarn } from './shared.js';
+import { inlineAllWasms } from './plugins/inline-all-wasms.js';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+export const apidomWorkerConfig = defineConfig({
+  configFile: false,
+  customLogger: logger,
+  mode: 'production',
+  plugins: [inlineAllWasms()],
+
+  build: {
+    lib: {
+      entry: resolve(__dirname, '../src/plugins/editor-monaco-language-apidom/language/apidom.worker.js'),
+      formats: ['iife'],
+      fileName: () => 'apidom.worker.js',
+      name: 'apidomWorker',
+    },
+    outDir: 'dist/umd',
+    sourcemap: false,
+    emptyOutDir: false,
+    codeSplitting: false,
+    rollupOptions: {
+      onwarn: sharedOnwarn,
+    },
+  },
+});
+
+export const editorWorkerConfig = defineConfig({
+  configFile: false,
+  customLogger: logger,
+  mode: 'production',
+  plugins: [],
+
+  build: {
+    lib: {
+      entry: resolve(__dirname, '../node_modules/monaco-editor/esm/vs/editor/editor.worker.js'),
+      formats: ['iife'],
+      fileName: () => 'editor.worker.js',
+      name: 'editorWorker',
+    },
+    outDir: 'dist/umd',
+    sourcemap: false,
+    emptyOutDir: false,
+    codeSplitting: false,
+    rollupOptions: {
+      onwarn: sharedOnwarn,
+    },
+  },
+});
